Refactoring app.tsx into smaller components
