+++ src/components/Sanitizer/Sanitizer.styles.ts
@@ -2,34 +2,31 @@ import { ICheckboxStyles, IStackStyles } from "@fluentui/react";
 
 const checkboxWidth = 300;
 
-export const logIssueLinkeStyle: React.CSSProperties = {
-  margin: '30px'
-}
-
 export const checkboxStyle: ICheckboxStyles = {
-  label: {
-    width: `${checkboxWidth}px`
-  }
+    label: {
+        width: `${checkboxWidth}px`
+    }
 }
 
 export const layoutStackStyle: React.CSSProperties = {
-  marginTop: '300px'
+    marginTop: '300px'
 }
 
+
 export const containerStyle: React.CSSProperties = {
-  border: '1px solid lightgray',
-  padding: '50px 130px',
-  borderRadius: '10px'
+    border: '1px solid lightgray',
+    padding: '50px 130px',
+    borderRadius: '10px'
 }
 
 export const radioButtonStackStyle: IStackStyles = {
-  root: {
-    width: `${checkboxWidth * 2 + 20}px`,
-    marginTop: '20px'
-  },
+    root: {
+        width: `${checkboxWidth * 2 + 20}px`,
+        marginTop: '20px'
+    },
 };
 
 export const fileUploadStyle: React.CSSProperties = {
-  marginTop: '50px',
-  display: 'inline-block'
+    marginTop: '50px',
+    display: 'inline-block'
 }
